+++ src/device/tsp.script/Script/FirmwareScriptBase.cs
@@ -1,4 +1,3 @@
-using System.Security.Cryptography;
 using System.Text;
 using cc.isr.Std.TrimExtensions;
 using cc.isr.VI.Syntax.Tsp;