+++ src/device/tsp/Tsp/Script/SessionBaseExtensions/Methods_save.cs
@@ -1,0 +1,35 @@
+using cc.isr.VI.Pith;
+
+namespace cc.isr.VI.Tsp.Script.SessionBaseExtensions;
+
+public static partial class Methods
+{
+    /// <summary>   A <see cref="SessionBase"/> extension method that saves a script to non-volatile-memory (NVM). </summary>
+    /// <remarks>   2025-04-10. </remarks>
+    /// <exception cref="ArgumentNullException">        Thrown when one or more required arguments
+    ///                                                 are null. </exception>
+    /// <exception cref="InvalidOperationException">    Thrown when the requested operation is
+    ///                                                 invalid. </exception>
+    /// <param name="session">      The session to act on. </param>
+    /// <param name="scriptName">   Name of the script. </param>
+    public static void SaveScript( this SessionBase session, string scriptName )
+    {
+        if ( session == null )
+            throw new ArgumentNullException( nameof( session ) );
+        if ( !session.IsSessionOpen )
+            throw new InvalidOperationException( $"{nameof( session )} is not open." );
+        if ( string.IsNullOrWhiteSpace( scriptName ) )
+            throw new ArgumentNullException( nameof( scriptName ) );
+
+        session.SetLastAction( $"checking if the {scriptName} script exists;. " );
+        session.LastNodeNumber = default;
+        if ( session.IsNil( scriptName ) )
+            throw new InvalidOperationException( $"The script {scriptName} cannot be run because it was not found." );
+
+        session.SetLastAction( $"saving the {scriptName} script;. " );
+        _ = session.WriteLine( $"{scriptName}.source=nil" );
+        _ = SessionBase.AsyncDelay( session.ReadAfterWriteDelay );
+    }
+
+
+}